+++ bidict/_abc.py
@@ -77,20 +77,6 @@ class BidirectionalMapping(_t.Mapping[KT, VT]):
         """
         return iter(self.inverse.items())
 
-    def values(self) -> _t.KeysView[VT]:  # type: ignore [override]  # https://github.com/python/typeshed/issues/4435
-        """A set-like object providing a view on the contained values.
-
-        Override the implementation inherited from
-        :class:`~collections.abc.Mapping`.
-        Because the values of a :class:`~bidict.BidirectionalMapping`
-        are the keys of its inverse,
-        this returns a :class:`~collections.abc.KeysView`
-        rather than a :class:`~collections.abc.ValuesView`,
-        which has the advantages of constant-time containment checks
-        and supporting set operations.
-        """
-        return self.inverse.keys()
-
 
 class MutableBidirectionalMapping(BidirectionalMapping[KT, VT], _t.MutableMapping[KT, VT]):
     """Abstract base class (ABC) for mutable bidirectional mapping types."""